--- next.config.js
@@ -1,16 +1,0 @@
-const path = require('path')
-
-module.exports = {
-  reactStrictMode: true,
-  sassOptions: {
-    includePaths: [path.join(__dirname, 'styles')]
-  },
-  webpack(config) {
-    config.module.rules.push({
-      test: /\.svg$/,
-      use: ['@svgr/webpack']
-    })
-
-    return config
-  }
-}
