+++ next.config.ts
@@ -1,0 +1,18 @@
+import type { NextConfig } from "next";
+const path = require('path')
+
+const nextConfig: NextConfig = {
+  sassOptions: {
+    includePaths: [path.join(__dirname, 'src', 'app', 'styles')]
+  },
+  webpack(config) {
+    config.module.rules.push({
+      test: /\.svg$/,
+      use: ['@svgr/webpack']
+    })
+
+    return config
+  }
+};
+
+export default nextConfig;
