+++ package.json
@@ -1,24 +1,28 @@
 {
-  "name": "portfolio-bruna",
+  "name": "portfolio-bruna-atualizado",
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev",
-    "build": "next build",
+    "dev": "next dev --turbopack",
+    "build": "next build --turbopack",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "eslint"
   },
   "dependencies": {
-    "next": "11.0.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2"
+    "next": "15.5.4",
+    "react": "19.1.0",
+    "react-dom": "19.1.0",
+    "sass": "^1.93.2"
   },
   "devDependencies": {
-    "@svgr/webpack": "^5.5.0",
-    "@types/react": "17.0.14",
-    "eslint": "7.31.0",
-    "eslint-config-next": "11.0.1",
-    "sass": "^1.35.2",
-    "typescript": "4.3.5"
+    "@eslint/eslintrc": "^3",
+    "@next/third-parties": "^15.5.4",
+    "@svgr/webpack": "^8.1.0",
+    "@types/node": "^20",
+    "@types/react": "^19",
+    "@types/react-dom": "^19",
+    "eslint": "^9",
+    "eslint-config-next": "15.5.4",
+    "typescript": "^5"
   }
 }
