+++ src/app/components/footer/styles.module.scss
@@ -1,0 +1,44 @@
+@use'variables.scss' as *;
+
+.footer {
+  background-color: #f9f9f9;
+  margin-top: 80px;
+  padding: 24px 0;
+
+  @media (min-width: $lg) {
+    padding: 30px 0;
+    margin-top: 170px;
+  }
+
+  > div {
+    align-items: center;
+    display: flex;
+    justify-content: space-between;
+  }
+}
+
+.year {
+  color: $text-01;
+  font-size: 12px;
+  line-height: 16px;
+  font-weight: bold;
+
+  @media (min-width: $lg) {
+    font-size: 14px;
+  }
+}
+
+.text {
+  color: $text-01;
+  font-size: 12px;
+  line-height: 16px;
+
+  @media (min-width: $lg) {
+    font-size: 14px;
+  }
+
+  a {
+    color: $primary;
+    font-weight: bold;
+  }
+}
