+++ src/app/components/go-to-top-button/index.tsx
@@ -1,0 +1,44 @@
+'use client'
+
+import { useEffect, useState } from 'react'
+import Image from 'next/image'
+
+import s from './styles.module.scss'
+
+const GoTotopButton = () => {
+  const [opacity, setOpacity] = useState(0)
+
+  useEffect(() => {
+    document.addEventListener('scroll', () => {
+      const scrollPosition = window.pageYOffset
+
+      if (scrollPosition <= window.innerHeight / 2) {
+        setOpacity(0)
+      } else {
+        setOpacity(1)
+      }
+    })
+  }, [])
+
+  const scrollToTop = () => {
+    window.scrollTo({
+      top: 0,
+      behavior: 'smooth'
+    })
+  }
+
+  return (
+    <div
+      className={`container ${s.wrapper}`}
+      style={{
+        opacity: opacity
+      }}
+    >
+      <button className={s.button} onClick={scrollToTop}>
+        <Image src="icons/arrow-button.svg" alt="arrow icon" width={47} height={47} />
+      </button>
+    </div>
+  )
+}
+
+export default GoTotopButton
