+++ src/app/components/go-to-top-button/styles.module.scss
@@ -1,0 +1,29 @@
+@use'variables.scss' as *;
+
+.wrapper {
+  bottom: 90px;
+  left: 50%;
+  position: fixed;
+  transform: translateX(-50%);
+  transition: opacity 0.3s ease;
+  width: 100%;
+  z-index: 0;
+
+  @media (min-width: $lg) {
+    bottom: 120px;
+  }
+}
+
+.button {
+  background-color: transparent;
+  bottom: 0;
+  border: none;
+  cursor: pointer;
+  right: 16px;
+  position: absolute;
+  z-index: 100;
+
+  @media (min-width: $lg) {
+    right: 0;
+  }
+}
