+++ src/app/styles/contact.module.scss
@@ -1,0 +1,17 @@
+@use'variables.scss' as *;
+
+.title {
+  color: $primary;
+}
+
+.content {
+  margin-top: 18px;
+
+  @media (min-width: $lg) {
+    margin-top: 25px;
+  }
+
+  p {
+    color: $text-02;
+  }
+}
