+++ src/app/styles/variables/breakpoints.scss
@@ -1,0 +1,1 @@
+$lg: 1200px;
