--- styles/about.module.scss
@@ -1,20 +1,0 @@
-@import 'variables.scss';
-
-.title {
-  color: $primary;
-}
-
-.content {
-  margin-top: 18px;
-
-  @media (min-width: $lg) {
-    margin-top: 25px;
-  }
-
-  p {
-    color: $text-02;
-    &:not(:first-of-type) {
-      margin-top: 24px;
-    }
-  }
-}
