--- styles/variables.scss
@@ -1,2 +1,0 @@
-@import './variables/breakpoints.scss';
-@import './variables/colors.scss';
