--- styles/variables/breakpoints.scss
@@ -1,1 +1,0 @@
-$lg: '1200px';
